// plugins/testkube/src/types/openapi.ts
export interface components {
  schemas: {
// ...
    TestWorkflowExecution: {
      /**
       * Format: bson objectId
       * @description unique execution identifier
       * @example 62f395e004109209b50edfc1
       */
      id: string;
      /**
       * Format: bson objectId
       * @description identifier for group of correlated executions
       * @example 62f395e004109209b50edfc1
       */
      groupId?: string;
      /** @description identifier of the runner where it has been executed */
      runnerId?: string;
      runnerTarget?: components['schemas']['ExecutionTarget'];
      runnerOriginalTarget?: components['schemas']['ExecutionTarget'];
      /**
       * @description execution name
       * @example some-workflow-name-1
       */
      name: string;
      /**
       * @description execution namespace
       * @example my-testkube
       */
      namespace?: string;
      /** @description sequence number for the execution */
      number?: number;
      /**
       * Format: date-time
       * @description when the execution has been scheduled to run
       */
      scheduledAt?: string;
      /**
       * Format: date-time
       * @description when the execution has been assigned to some runner
       */
      assignedAt?: string;
      /**
       * Format: date-time
       * @description when the execution result's status has changed last time (queued, passed, failed)
       */
      statusAt?: string;
      /** @description structured tree of steps */
      signature?: components['schemas']['TestWorkflowSignature'][];
      result: components['schemas']['TestWorkflowResult'];
      /** @description additional information from the steps, like referenced executed tests or artifacts */
      output?: components['schemas']['TestWorkflowOutput'][];
      /** @description generated reports from the steps, like junit */
      reports?: components['schemas']['TestWorkflowReport'][];
      resourceAggregations?: components['schemas']['TestWorkflowExecutionResourceAggregationsReport'];
      workflow: components['schemas']['TestWorkflow'];
      resolvedWorkflow?: components['schemas']['TestWorkflow'];
      /** @description test workflow execution name started the test workflow execution */
      testWorkflowExecutionName?: string;
      /**
       * @description whether webhooks on the execution of this test workflow are disabled
       * @default false
       * @example [
       *       true,
       *       false
       *     ]
       */
      disableWebhooks: boolean;
      tags?: components['schemas']['TestWorkflowTagValue'];
      /** @description running context for the test workflow execution (Pro edition only) */
      runningContext?: components['schemas']['TestWorkflowRunningContext'];
      configParams?: components['schemas']['TestWorkflowExecutionConfig'];
    };
// ...
  };
// ...
}